Support showing multiple years
Add support to show at least 5 FOSDEM editions by checking the current
date and if FOSDEM already happend this year. Now since I first always
fetched $current year we now have to keep track of the current year in
every route so that it can fetch the correct data and fetch the relevant
information. This is subpar and should be improved either by passing XML
or parsing the required information and transforming it already.